feat：本地和Docker支持交互式的用户代码

+++ src/main/java/com/zyf/zojcodesandbox/model/ExecuteMessage.java
@@ -15,5 +15,5 @@ public class ExecuteMessage {
 
     private Long time;
 
-    private Long memory;
+    private Double memory;
 }
